LayoutTests/imported/w3c/web-platform-tests/cookiestore/cookieStore_get_set_across_origins.sub.https-expected.txt

CONSOLE MESSAGE: Blocked access to external URL https://www1.localhost:9443/cookiestore/resources/helper_iframe.sub.html
CONSOLE MESSAGE: Unable to post message to https://www1.localhost:9443. Recipient has origin https://localhost:9443.


Harness Error (TIMEOUT), message = null

TIMEOUT cookieStore.get() sees cookieStore.set() in cross-origin frame Test timed out
NOTRUN cookieStore.get() in cross-origin frame sees cookieStore.set()
NOTRUN cookieStore.set() in cross-origin does not overwrite the __Host- cookie
NOTRUN __Host- cookies set via cookieStore.set() in same-site domains don't overwrite each other
